SuivA.Windows.Tests/UnitTest/ActionCommandTests.cs
using System;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace SuivA.Windows.Tests.UnitTest
{
    [TestClass]
    public class ActionCommandTests
    {
        [TestMethod]
        [ExpectedException(typeof(ArgumentNullException))]
        public void ConstructorThrowsExceptionIfActionParameterIsNull()
        {
            var unused = new ActionCommand(null);
        }

        [TestMethod]
        public void ExecuteInvokesAction()
        {
            var invoked = false;

            void Action(object obj)
            {
                invoked = true;
            }

            var command = new ActionCommand(Action);

            command.Execute();

            Assert.IsTrue(invoked);
        }

        [TestMethod]
        public void ExecuteOverloadInvokesActionWithParameter()
        {
            var invoked = false;

            void Action(object obj)
            {
                Assert.IsNotNull(obj);
                invoked = true;
            }

            var command = new ActionCommand(Action);

            command.Execute(new object());

            Assert.IsTrue(invoked);
        }

        [TestMethod]
        public void CanExecuteIsTrueByDefault()
        {
            var command = new ActionCommand(obj => { });
            Assert.IsTrue(command.CanExecute(null));
        }

        [TestMethod]
        public void CanExecuteOverloadExecutesTruePredicate()
        {
            var command = new ActionCommand(obj => { }, obj => (int) obj == 1);
            Assert.IsTrue(command.CanExecute(1));
        }

        [TestMethod]
        public void CanExecuteOverloadExecutesFalsePredicate()
        {
            var command = new ActionCommand(obj => { }, obj => (int) obj == 1);
            Assert.IsFalse(command.CanExecute(0));
        }
    }
}